requirements.txt
groq
pydantic
langchain-text-splitters
qdrant-client
sentence-transformers
python-dotenv
langgraph
langsmith
fastapi
uvicorn
prometheus-client
python-multipart
PyMuPDF
python-docx
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
bcrypt
